requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
hvac
motor
jinja2>=3.0.0
python-multipart
#added for monitoring
prometheus-client==0.20.0
starlette-exporter==0.21.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-exporter-otlp==1.26.0
